# Proyecto/E3 (Final)/BACKEND/swagger.yaml
openapi: 3.0.0
info:
  title: Dawdle API
  description: |
    API para gestión de eventos y usuarios
    
    ## Autenticación
    La API utiliza un sistema híbrido de autenticación:
    - **JWT (JSON Web Token)**: Se devuelve en el cuerpo de la respuesta con expiración de 1 hora
    - **Cookies httpOnly**: Se establece automáticamente en las rutas `/auth/register` y `/auth/login` con las siguientes características:
      - `httpOnly: true` - No accesible desde JavaScript (protección contra XSS)
      - `secure: true` - Solo en HTTPS (producción)
      - `sameSite: 'lax'` - Protección contra CSRF
      - `maxAge: 3600000` - 1 hora (3600 segundos)
      - `path: '/'` - Disponible en toda la aplicación
    
    ## Seguridad de contraseñas
    Las contraseñas se almacenan utilizando bcrypt con **12 rounds de salt** para máxima seguridad.
    
    ## CORS
    El servidor acepta peticiones solo desde el origen configurado en `FRONT_ORIGIN` (variable de entorno).
  version: 1.0.0
  contact:
    name: Soporte Dawdle
    email: soporte@dawdle.com
servers:
  - url: http://localhost:3000
    description: Servidor local
  - url: https://dawdle-backend.onrender.com
    description: Servidor de producción
...
paths:
  /auth/register:
    post:
      tags:
        - Auth
      summary: Registrar nuevo usuario
      description: Crea una nueva cuenta de usuario
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - username
                - email
                - password
                - age
                - birthday
              properties:
                username:
                  type: string
                  example: "nuevo_usuario"
                email:
                  type: string
                  format: email
                  example: "usuario@example.com"
                password:
                  type: string
                  format: password
                  example: "MiPassword123"
                age:
                  type: integer
                  example: 25
                birthday:
                  type: string
                  format: date
                  example: "1999-05-20"
      responses:
        '200':
          description: Registro exitoso. Se establece cookie httpOnly con el token (maxAge 1h, sameSite lax, secure en producción)
          headers:
            Set-Cookie:
              description: Cookie de autenticación (httpOnly)
              schema:
                type: string
                example: "authToken=eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...; Path=/; HttpOnly; SameSite=Lax; Max-Age=3600"
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Usuario creado exitosamente"
                  user:
                    $ref: '#/components/schemas/User'
                  token:
                    type: string
                    description: JWT con expiración de 1 hora
                    example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
        '400':
          description: Error de validación en los datos enviados
          content:
            application/json:
              schema:
                type: object
                properties:
                  errors:
                    type: array
                    items:
                      type: object
                      properties:
                        msg:
                          type: string
                        param:
                          type: string
                example:
                  errors:
                    - msg: "El username debe tener entre 3 y 50 caracteres"
                      param: "username"
        '403':
          description: Username reservado
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "El username 'admin' está reservado y no puede ser registrado"
        '409':
          description: Usuario o email ya existe
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                example:
                  error: "El email ya está registrado"
  /auth/login:
    post:
      tags:
        - Auth
      summary: Iniciar sesión
      description: Autenticar usuario con email y password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
              properties:
                email:
                  type: string
                  format: email
                  example: "usuario@example.com"
                password:
                  type: string
                  format: password
                  example: "MiPassword123"
      responses:
        '200':
          description: Inicio de sesión exitoso. Se establece cookie httpOnly con el token (maxAge 1h, sameSite lax, secure en producción)
          headers:
            Set-Cookie:
              description: Cookie de autenticación (httpOnly)
              schema:
                type: string
                example: "authToken=eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...; Path=/; HttpOnly; SameSite=Lax; Max-Age=3600"
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Inicio de sesión exitoso"
                  user:
                    $ref: '#/components/schemas/User'
                  token:
                    type: string
                    description: JWT con expiración de 1 hora. También se envía en cookie httpOnly para mayor seguridad
                    example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
        '400':
          description: Error de validación
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  errors:
                    type: array
                example:
                  message: "Error al iniciar sesión"
                  errors:
                    - msg: "Email inválido"
                      param: "email"
        '401':
          description: Credenciales inválidas
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Credenciales inválidas"
...
  /auth/me:
    get:
      tags:
        - Auth
      summary: Obtener información del usuario actual
      description: Retorna los datos del usuario autenticado
      responses:
        '200':
          description: Información del usuario obtenida
          content:
            application/json:
              schema:
                type: object
                properties:
                  user:
                    $ref: '#/components/schemas/User'
        '401':
          description: Token no proporcionado, inválido o expirado
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                example:
                  error: "Acceso denegado. Token no proporcionado."
...
components:
  schemas:
    User:
      type: object
      properties:
        id:
          type: integer
          example: 1
        username:
          type: string
          example: "nuevo_usuario"
        email:
          type: string
          format: email
          example: "usuario@example.com"
        image:
          type: string
          format: uri
          example: "https://ui-avatars.com/api/?name=nuevo_usuario&background=random"
        age:
          type: integer
          example: 25
        birthday:
          type: string
          format: date-time
          example: "1999-05-20T00:00:00.000Z"
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: []